pyproject.toml
[project]
name = "knitting-machine"
version = "0.1.0"
description = "Tricking a Brother knitting machine into thinking that a computer is a disk drive"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pillow>=12.1.0",
    "pyserial>=3.5",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "python-multipart>=0.0.9",
]

[tool.setuptools]
packages = ["app"]

[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[dependency-groups]
dev = [
    "black>=26.1.0",
    "flake8>=7.3.0",
    "mypy>=1.19.1",
    "types-pyserial>=3.5.0.20260408",
    "pytest>=8.0",
    "httpx",
]

[tool.black]
line-length = 88
target-version = ['py313']
skip-string-normalization = false

[tool.mypy]
explicit_package_bases = true
mypy_path = "tests"
